Conflicts:
	trg-s3.cc

+++ trg-s3.cc
@@ -549,7 +549,7 @@ void DoLoopContraction(DTensor14 *C, const DTensor9 &K, const DTensor9 &SU,
 }
 
 void TRGS3(const double a, const double b, const double c,
-    const uint8_t dc, const double condi, const uint8_t iter,
+    const uint8_t dc, const double condi, const unsigned iter,
     const DTensor9 &K) {
   std::cout << K << std::endl;
   DTensor4 B1;
@@ -575,7 +575,7 @@ void TRGS3(const double a, const double b, const double c,
   MakeSecondBlocks(&B2, m_A, m_B, &C1, ind, sv_len, rho_A, rho_B, condi);
   std::cout << B2 << std::endl;
 
-  for (uint8_t i = 0; i < iter; ++i) {
+  for (unsigned i = 0; i < iter; ++i) {
     DTensor9 SVD2[2];
     DoLoopSVD(SVD2, sv_len, &B2, dc, condi, rho_A, rho_B, m_A, m_B);
     B2.Clear();
